5.0.4 remove script on unmount
