+++ demo/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
-  <link rel="stylesheet" href="//cdn.rawgit.com/twbs/bootstrap/v4-dev/dist/css/bootstrap.css">
-  <div id="google-login"></div>
+  <link rel="stylesheet" href="//cdn.rawgit.com/twbs/bootstrap/v4-dev/dist/css/bootstrap.css" />
+  <div id="app"></div>
   <script src="//use.fontawesome.com/3bd7769ce1.js"></script>
   <script src="/webpack-dev-server.js"></script>
   <script src="bundle.js"></script>
