+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-google-login",
-  "version": "5.0.3",
+  "version": "5.0.4",
   "description": "A Google Login Component for React",
   "main": "dist/google-login.js",
   "scripts": {
@@ -11,7 +11,7 @@
     "test:watch": "jest --watch",
     "lint": "eslint ./src",
     "lint:fix": "eslint ./src --fix",
-    "clean": "rm -rf node_modules dist; rm package-lock.json"
+    "fresh": "rm -rf node_modules dist; rm package-lock.json; npm i"
   },
   "pre-commit": [
     "test",
@@ -44,32 +44,32 @@
     "prop-types": "^15.6.0"
   },
   "devDependencies": {
-    "@babel/core": "7.2.2",
-    "@babel/plugin-proposal-object-rest-spread": "7.2.0",
-    "@babel/plugin-transform-runtime": "7.2.0",
-    "@babel/preset-env": "7.2.3",
+    "@babel/core": "7.4.3",
+    "@babel/plugin-proposal-object-rest-spread": "7.4.3",
+    "@babel/plugin-transform-runtime": "7.4.3",
+    "@babel/preset-env": "7.4.3",
     "@babel/preset-react": "7.0.0",
-    "autoprefixer": "9.4.5",
+    "autoprefixer": "9.5.0",
     "babel-core": "7.0.0-bridge.0",
-    "babel-jest": "23.6.0",
+    "babel-jest": "24.7.1",
     "babel-loader": "8.0.5",
     "babel-plugin-transform-react-constant-elements": "6.23.0",
     "babel-plugin-transform-react-inline-elements": "6.22.0",
-    "babel-plugin-transform-react-remove-prop-types": "0.4.21",
-    "enzyme": "3.8.0",
-    "enzyme-adapter-react-16": "1.7.1",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.24",
+    "enzyme": "3.9.0",
+    "enzyme-adapter-react-16": "1.11.2",
     "eslint-config-ag": "2.0.0",
-    "jest": "23.6.0",
+    "jest": "24.7.1",
     "pre-commit": "^1.2.2",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-fontawesome": "1.6.1",
-    "react-hot-loader": "4.6.3",
-    "react-test-renderer": "16.7.0",
-    "uglifyjs-webpack-plugin": "2.1.1",
-    "webpack": "4.28.4",
-    "webpack-cli": "3.2.1",
-    "webpack-dev-server": "3.1.14"
+    "react-hot-loader": "4.8.2",
+    "react-test-renderer": "16.8.6",
+    "uglifyjs-webpack-plugin": "2.1.2",
+    "webpack": "4.29.6",
+    "webpack-cli": "3.3.0",
+    "webpack-dev-server": "3.2.1"
   },
   "peerDependencies": {
     "react": "^16.0.0",
